+++ Simulated_annealing/Structures.h
@@ -13,9 +13,9 @@
 
 
 struct InputDate{
-    int proc_num = 0;
-    int work_amount = 0;
-    std::vector <int> data = {};
+    size_t proc_num = 0;
+    size_t work_amount = 0;
+    std::vector <size_t> data = {};
 };
 
 class SingleLoad {
@@ -133,8 +133,6 @@ public:
 
     }
 
-
-
     virtual bool EmptyLine(int position) {
         return global_loading[position].Empty();
     }
@@ -143,7 +141,6 @@ public:
         std::cout << "Answer:" << std::endl;
         for (size_t i=0; i<amount; i++) {
             std::cout << "Proc №"<< i << " --- " << global_loading[i].GetLen() << std::endl;
-//            global_loading[i].Out();
         }
         std::cout << "Deviation - " << this->CriterionGet() << std::endl;
     }
